@@ -1,20 +1,24 @@
 {
-    "name": "volcanus/thumbnail",
-    "homepage": "https://github.com/k-holy/volcanus-thumbnail",
-    "type": "library",
-    "version": "0.7.1",
-    "license": "MIT",
-    "authors": [
-        {
-            "name": "k-holy",
-            "email": "k.holy74@gmail.com"
-        }
-    ],
-    "autoload": {
-        "psr-4": { "Volcanus\\Thumbnail\\": "src/" }
-    },
-    "require": {
-        "php": ">=5.5.9",
-        "ext-mcrypt" : "*"
+  "name": "volcanus/thumbnail",
+  "homepage": "https://github.com/k-holy/volcanus-thumbnail",
+  "type": "library",
+  "version": "0.7.1",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "k-holy",
+      "email": "k.holy74@gmail.com"
     }
+  ],
+  "autoload": {
+    "psr-4": {
+      "Volcanus\\Thumbnail\\": "src/"
+    }
+  },
+  "require": {
+    "php": "^5.6||^7.0||^7.1"
+  },
+  "require-dev": {
+    "phpunit/phpunit": "^5.7.20||^6.0"
+  }
 }
