@@ -2,10 +2,11 @@
 /**
  * Volcanus libraries for PHP
  *
- * @copyright 2011-2013 k-holy <k.holy74@gmail.com>
+ * @copyright k-holy <k.holy74@gmail.com>
  * @license The MIT License (MIT)
  */
 error_reporting(E_ALL | E_STRICT | E_DEPRECATED);
 
+/** @noinspection PhpIncludeInspection */
 $loader = include realpath(__DIR__ . '/../vendor/autoload.php');
 $loader->add('Volcanus\Thumbnail\Test', __DIR__);
